Feat: Completes Frontend for CA Leaderboards
